Check if the owner already has a restaurant and if yes load it

+++ app/src/main/java/com/example/food2you_restaurantsonly/data/local/entities/Restaurant.kt
@@ -11,7 +11,7 @@ data class Restaurant(
     val kitchen: String,
     val deliveryPrice: Float,
     val deliveryTimeMinutes: Int,
-    val minimalPrice: Int,
+    val minimalPrice: Float,
     val imgUrl: String,
     val previews: List<String>,
     val users: List<String>,
